Add literals (string, array, dictionary) and these accessors. Delegate the stack of the Environment class to the Stack class. 1. 文字列、配列、辞書サポート     1. opArray, opDictionary, opIndex     2. JpfString.emit(with:)追加     3. GenitiveExpression.compiled(with:)を修正         1. rightがIdentifierで、leftがArrayLiteralまたはDictionrayLiteralまたはGenitiveExpressionの場合、コンパイル         2. rightがPredicateExpressionで、leftがIdentifierの場合、コンパイル     4. StringLiteral.compiled(with:)、ArrayLiteral.compiled(with:)、DictionaryLiteral.compiled(with:)を追加     5. GenitiveExecuter.execute()を追加 2. Environmentのスタック([JpfObject])をclass Stackに変更(処理を委譲)     1. Environment.init(with stack: Stack)を追加     2. Environment.push(_:), Compiler.push(_:)の返り値をJpfError?に変更     3. Stack.pull()の返り値をJpfObjectからJpfObject?に変更(取り出す値が無ければnil)     4. Stack.peek(_:)で、個数が0の場合、[]を返す     5. Stack.drop(_:)で、個数がcount以上であれば、count分捨てる。     6. StackクラスをEnvironmentに対応・拡張 3. 修正     1. Environment.apply(_:numberOfArguments:from:)をリファクタリング     2. Array<JpfObject>.stringを追加     3. Replで、翻訳後の定数表(constans.string)を表示     4. AddExecuter.execute()をAddOperator.operated()を使って実行(文字列の足し算や、可変長項の足し算に対応)

+++ JPForce/evaluatable.swift
@@ -68,8 +68,10 @@ extension ExpressionStatement : Evaluatable {
         var result: JpfObject?
         for expression in expressions {
             result = expression.evaluated(with: environment)
-            if let object = result, object.isBreakFactor {break}
-            result.map {environment.push($0)}
+            if let object = result {
+                if object.isBreakFactor {break}
+                if let err = environment.push(object) {return err}
+            }
         }
         environment.remove(name: Environment.OUTER)
         return result
@@ -225,7 +227,7 @@ extension Boolean : Evaluatable {
 }
 extension RangeLiteral : Evaluatable {
     func evaluated(with environment: Environment) -> JpfObject? {
-        environment.push(JpfNull.object)    // 範囲の上下限が、配列にアクセスしないための回避策
+        if let err = environment.push(JpfNull.object) {return err}  // 範囲の上下限が、配列にアクセスしないための回避策
         var lowerBound, upperBound: (JpfInteger, Token)?
         if let object = self.lowerBound?.evaluated(with: environment) {
             if object.isError {return object}
@@ -554,8 +556,10 @@ extension LoopExpression : Evaluatable {
         var result: JpfObject?
         for expression in condition {
             result = expression.evaluated(with: environment)
-            if let object = result, object.isBreakFactor {break}
-            result.map {environment.push($0)}
+            if let object = result {
+                if object.isBreakFactor {break}
+                if let err = environment.push(object) {return err}
+            }
         }
         environment.drop()                          // 真偽値を捨てる
         return result
@@ -856,7 +860,7 @@ extension GenitiveExpression : Evaluatable {
         let phrase = JpfPhrase(name: object.name, value: object, particle: token)
         switch right {
         case let ident as Identifier:
-            environment.push(phrase)
+            if let err = environment.push(phrase) {return err}
             accessor = ident.evaluated(with: environment)
             if environment.peek?.name == object.name && environment.isPeekParticle(.NO) {
                 environment.drop()
@@ -864,7 +868,7 @@ extension GenitiveExpression : Evaluatable {
                 return accessor
             }
         case is ComputationLiteral:
-            environment.push(phrase)
+            if let err = environment.push(phrase) {return err}
             if let c = right.evaluated(with: environment) as? JpfComputation {
                 return environment.isExecutable ? c.getter(with: environment) : c
             }
@@ -873,7 +877,7 @@ extension GenitiveExpression : Evaluatable {
             let object = evaluated(object, expression.left, with: environment)
             return JpfPhrase(name: "", value: object, particle: expression.token)
         case is PredicateExpression, is CaseExpression, is Label:
-            environment.push(phrase)
+            if let err = environment.push(phrase) {return err}
             return right.evaluated(with: environment)
         default:
             accessor = right.evaluated(with: environment)
